// packages/cosapp-lab/cosapp_lab-0.15.0.tar.gz/cosapp_lab-0.15.0/src/sys_explorer_react/redux/types.ts
export interface UserListType {
  uid: number;
  name: string;
}